# Better JpegEncoder profiling & benchmarks