@JimBobSquarePants 2021-02-05
Love it @antonfirsov this is great insight!

>This means that our current primary bottleneck is Huffmann encoding.

I thought it would be the case. I was reading up on a Cloudflare blog post talking about optimising jpegtran Huffman encoding using SIMD the other day plus I spotted that libjpeg turbo also has specific SSE versions.