@antonfirsov 2021-02-07
That `>10%` actually feels very surprising to me. I'm actually worried that either there is something wrong with the benchmark, or we are doing something really-really wrong.

Anyways, found the Cloudfare article mentioned:
https://blog.cloudflare.com/doubling-the-speed-of-jpegtran/

There's also a good general article by Intel:
https://software.intel.com/content/www/us/en/develop/articles/fast-computation-of-huffman-codes.html

Also notice the `Stream.Write` bottleneck. We could use some double buffering probably just like in the decoder.