@JimBobSquarePants 2021-02-08
I know Mango have been working in this area also. Perhaps they have some good ideas. 
https://github.com/t0rakka/mango